- 4.22 blogs

+++ controllers/blogs.js
@@ -2,6 +2,8 @@ const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
 const jwt = require('jsonwebtoken');
+const middleware = require('../utils/middleware');
+const userExtractor = middleware.userExtractor;
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {
@@ -12,7 +14,13 @@ blogsRouter.get('/', async (request, response) => {
 });
 
 blogsRouter.get('/:id', async (request, response, next) => {
-  const blog = await Blog.findById(request.params.id);
+  const blog = await Blog.findById(request.params.id).populate(
+    'user',
+    {
+      username: 1,
+      name: 1,
+    }
+  );
   if (blog) {
     response.json(blog);
   } else {
@@ -20,42 +28,43 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 });
 
-blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body;
-  if (!body.title) {
-    response.status(400).end();
+blogsRouter.post(
+  '/',
+  userExtractor,
+  async (request, response, next) => {
+    const body = request.body;
+    if (!body.title) {
+      response.status(400).end();
+    }
+    const user = await request.user;
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user._id,
+    });
+    try {
+      const savedBlog = await blog.save();
+      user.blogs = user.blogs.concat(savedBlog._id);
+      await user.save();
+      response.status(201).json(savedBlog);
+    } catch (e) {
+      next(e);
+    }
   }
-  const token = request.token;
-  let decodedToken = {};
-  if (token) {
-    decodedToken = jwt.verify(token, process.env.SECRET);
-  }
-  if (!token || !decodedToken?.id) {
-    return response
+);
+
+blogsRouter.delete('/:id', userExtractor, async (req, res) => {
+  const user = await req.user;
+  const blog = await Blog.findById(req.params.id);
+
+  if (!(blog.user.toString() === user._id.toString())) {
+    return res
       .status(401)
-      .json({ error: 'token missing or invalid' });
+      .json({ error: 'You are not the blog creator' });
   }
-
-  const user = await User.findById(decodedToken.id);
-
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
-    user: user._id,
-  });
-  try {
-    const savedBlog = await blog.save();
-    user.blogs = user.blogs.concat(savedBlog._id);
-    await user.save();
-    response.status(201).json(savedBlog);
-  } catch (e) {
-    next(e);
-  }
-});
-
-blogsRouter.delete('/:id', async (req, res) => {
   await Blog.findByIdAndRemove(req.params.id);
   res.status(204).end();
 });
